backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
librosa==0.10.2
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1
ffmpeg-python==0.2.0
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
aiofiles==24.1.0
